UBUNTU: SAUCE: Return correct error code for mediated network connections
Fix AppArmor portion of Bug #647071 which is a bug in the ubuntu sauce
AppArmor network compatibility patch.  All network mediation was failing
to be applied because the correct error code was not being set in the
audit structure resulting in a return code of 0 always being returned.

BugLink: http://bugs.launchpad.net/bugs/647071

Signed-off-by: John Johansen <john.johansen@canonical.com>
Acked-by: Steve Conklin <steve.conklin@canonical.com>
Signed-off-by: Leann Ogasawara <leann.ogasawara@canonical.com>