SRU: UBUNTU: [Config] Disable inte_idle for -virtual kernels
BugLink: http://bugs.launchpad.net/bugs/651370

EC2 Xen does not properly handle the mwait instruction used by intel_idle
resulting in an invalid opcode when an instance is launched on newer
intel hardware.

SRU Justification:

Impact: Booting an Intel based instance with certain CPU level will fail with a panic as the driver does not seem to take into account that it is running in a virtualized environment. This only is a problem with the intel_idle driver.

Fix: Turning off intel_idle driver support for the virtual kernel image will let it use the generic idle driver as before. As this option is only changed for the virtual kernel package there is no risk of regression for the generic packages.

Testcase: Booting a large instance (with 68GB of memory) very likely results in this panic as the memory size will result in selecting certain base hardware with Intel CPUs. Turning the option off lets those instances boot again.

Signed-off-by: John Johansen <john.johansen@canonical.com>
Acked-by: Leann Ogasawara <leann.ogasawara@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Brad Figg <brad.figg@canonical.com>