UBUNTU: Temporarily disable abi checking
The omap flavour is failing to build from source due to a missing
module. The ignore.modules file was checked in empty which caused
the problem.

However, since the previous upload was an abi bump, it's not as
simple as renaming the abi tree and fixing the ignore.modules file.
That just results in the abi check failing.

This change fixes the ignore.modules file as well as ignoreing the
abi check for this next upload.

Signed-off-by: Brad Figg <brad.figg@canonical.com>